chromium/src/third_party/WebKit/LayoutTests/http/tests/cowl/xhr/response-labeled-json.html

<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

var test = async_test("Test XMLHttpRequest with responseType set to 'labeled-json' expecting success.");

test.step(function() {
    var lobj = new LabeledObject('password', {confidentiality: new Label(window.location.origin)});
    var xhr = new XMLHttpRequest;
    xhr.responseType = 'labeled-json';
    xhr.open('POST', 'resources/post-echo.php', true);
    var handler = function(test, json) {
        if (this.readyState != 4)
            return;

        assert_equals(this.status, 200, 'xhr.status');
        assert_equals(this.response.protectedObject, 'password', 'xhr.response.protectedObject');
        assert_equals(this.response.confidentiality.toString(), window.location.origin, 'xhr.response.confidentiality');
        assert_equals(this.response.integrity.toString(), "'none'", 'xhr.response.integrity');

        test.done();
    };
    xhr.onreadystatechange = test.step_func(handler.bind(xhr, test, lobj));
    xhr.send(lobj);
});

</script>
</body>
